requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
